+++ engine/src/com/twinsoft/convertigo/engine/admin/services/studio/dbo/Move.java
@@ -35,7 +35,7 @@ import com.twinsoft.convertigo.engine.admin.services.studio.ngxbuilder.BuilderUt
 import com.twinsoft.convertigo.engine.enums.FolderType;
 import com.twinsoft.convertigo.engine.flow.FlowStudioSupport;
 
-@ServiceDefinition(name = "Move", roles = { Role.WEB_ADMIN, Role.PROJECT_DBO_VIEW }, parameters = {}, returnValue = "")
+@ServiceDefinition(name = "Move", roles = { Role.WEB_ADMIN, Role.PROJECT_DBO_CONFIG }, parameters = {}, returnValue = "")
 public class Move extends JSonService {
 
 	@Override