+++ engine/src/com/twinsoft/convertigo/engine/translators/DefaultServletTranslator.java
@@ -52,14 +52,12 @@ public class DefaultServletTranslator implements Translator {
 //			parameterNames = reparsedParameters.keys();
 //		}
 		
-		boolean isHandleComplex = "true".equals(request.getParameter("__handleComplex"));
-		
 		while (parameterNames.hasMoreElements()) {
 			String parameterName = (String) parameterNames.nextElement();
 			String[] parameterValues = request.getParameterValues(parameterName);
 
 			if (!inputDocumentBuilder.handleSpecialParameter(parameterName, parameterValues)) {
-				inputDocumentBuilder.addVariable(parameterName, parameterValues, isHandleComplex);
+				inputDocumentBuilder.addVariable(parameterName, parameterValues);
 			}
 		}
 		
@@ -73,7 +71,7 @@ public class DefaultServletTranslator implements Translator {
 							((XMLVector<?>) parameterObject).toArray(new String[0]) : new String[] {(String) parameterObject};
 	
 					if (!inputDocumentBuilder.handleSpecialParameter(parameterName, parameterValues)) {
-						inputDocumentBuilder.addVariable(parameterName, parameterValues, isHandleComplex);
+						inputDocumentBuilder.addVariable(parameterName, parameterValues);
 					}
 				}
 			}